change dir structure
