+++ server/src/entities/board.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 use std::error::Error;
-use crate::state::{Room, Board};
+use crate::libs::state::{Room, Board};
 use tokio_postgres::Client;
