+++ server/src/lifecycle/mod.rs
@@ -1,0 +1,3 @@
+mod cleanup;
+
+pub use cleanup::remove_unused_rooms;
