+++ server/src/websocket/message.rs
@@ -4,8 +4,9 @@ use tokio_postgres::Client;
 use warp::ws::Message;
 use serde::{Deserialize, Serialize};
 use weak_table::WeakHashSet;
-use crate::{state::{Room, BoardSize, Command, CommandName}, board::get};
-use super::state::{Rooms, WSUsers, PullData};
+use crate::libs::state::{Room, BoardSize, Command, CommandName};
+use crate::entities::board::get;
+use crate::libs::state::{Rooms, WSUsers, PullData};
 use std::{mem, collections::HashMap, sync::Arc};
