+++ server/src/websocket/mod.rs
@@ -1,0 +1,4 @@
+mod connect;
+mod message;
+
+pub use connect::user_connected;
